feat: get rid of save_errors flag to simplify things
It defaults to `true` in the Wayback Machine GitHub Action anyway,
which is what I think makes the most sense. I can't come up with a
reason why users would want to set it to `false` in the context of
Waystation. I'm getting rid of it to make things simpler.